Refactor MemberCards with MemberCardsContainer
