+++ src/containers/index.js
@@ -1,0 +1,3 @@
+import MemberCardsContainer from './MemberCardsContainer'
+
+export { MemberCardsContainer }
